Only admin can have access of website

+++ myproject/subcat/views.py
@@ -6,22 +6,51 @@ from cat.models import Cat
 
 def subcat_list(request):
 
+    #Login Check Start
+    if not request.user.is_authenticated :
+        return redirect('my_login')
+    #Login Check End 
+
+    #If user is master user or admin then hes allowed to access this section
+    #Users who doesnt belong to this group are not authorised to this section
+    #CHECK START
+    perm = 0 
+    for i in request.user.groups.all(): 
+        if i.name == "masteruser" : perm = 1 
+
+    #CHECK END   
+
     subcat = SubCat.objects.all()
 
 
-    return render(request , 'back/subcat_list.html' , {'subcat' : subcat})
+    return render(request , 'back/subcat_list.html' , {'subcat' : subcat , 'perm' : perm})
 
 
 def subcat_add(request):
 
+    #Login Check Start
+    if not request.user.is_authenticated :
+        return redirect('my_login')
+    #Login Check End 
+
+    #If user is master user or admin then hes allowed to access this section
+    #Users who doesnt belong to this group are not authorised to this section
+    #CHECK START
+    perm = 0 
+    for i in request.user.groups.all(): 
+        if i.name == "masteruser" : perm = 1 
+
+    #CHECK END   
+
     cat = Cat.objects.all()
 
+    #Get all the entered data
     if request.method == 'POST' : 
         name = request.POST.get('name')
         catid = request.POST.get('cat')
 
+        #Error Page
         if name == "" :
-
             error = "All Fields Required"
             return render(request , 'back/error.html' , {'error' : error}) 
 
@@ -32,15 +61,21 @@ def subcat_add(request):
 
         catname = Cat.objects.get(pk = catid).name 
 
+        #Store it in the database
         b = SubCat(name = name , catname = catname , catid = catid)
         b.save()
 
         return redirect('subcat_list')
-    return render(request , 'back/subcat_add.html' , {'cat' : cat})
+    return render(request , 'back/subcat_add.html' , {'cat' : cat , 'perm' : perm})
 
 
 #Function to delete subcat from the Panel
 def subcat_del(request , pk):
+
+    #Login Check Start
+    if not request.user.is_authenticated :
+        return redirect('my_login')
+    #Login Check End 
 
     b = SubCat.objects.filter(pk=pk)
     b.delete()
